Add files via upload
This for Pavo Pico Version 1 - as of Wednesday October 25th it's working fine. Watch video for additional setting that may need to be added in goggles so DJI remote controller binds. 

Open txt file
Copy text
Go in to Betaflight
Go into the CLI past this in
Save

Watch video for addition goggle settings.